🔵 moved shared types to globals
